requirements.txt
ajax==3.0.0
aniso8601==9.0.1
asgiref==3.5.2
backports.entry-points-selectable==1.1.0
blinker==1.5
browser==0.0.1
brython==3.10.6
CacheControl==0.12.6
cachy==0.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cleo==0.8.1
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.8
cssselect==1.0.3
dacite==1.6.0
dataclasses-json==0.5.7
DateTime==4.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
Django==4.1
django-appconf==1.0.5
dsmtpd==0.3
EasyProcess==1.1
filelock==3.0.12
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
fonttools==4.37.2
fpdf2==2.5.7
html5lib==1.1
idna==3.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==21.8.0
lockfile==0.12.2
lxml==4.5.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==9.2.0
pipenv==2021.5.29
pkginfo==1.7.1
platformdirs==2.3.0
poetry==1.1.8
poetry-core==1.0.4
ptyprocess==0.7.0
pycparser==2.20
PyJWT==2.4.0
pylev==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyVirtualDisplay==0.2.4
requests==2.26.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selenium==3.141.0
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
svg.path==6.2
tomlkit==0.7.2
typing-inspect==0.8.0
typing_extensions==4.3.0
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.8.1
zope.interface==5.4.0
